+++ tests/conftest.py
@@ -7,11 +7,12 @@ import pytest
 import pytest_asyncio
 from pkg_resources import resource_filename
 
-from omc4py import latest, open_session
+from omc4py import open_session
 from omc4py.interactive import Interactive
 from omc4py.protocol import asynchronous
+from omc4py.v_1_21 import AsyncSession, Session  # NOTE: update to latest
 
-from .session.aio import EmptySession, NestedSession, OneSession
+from .session import AsyncEmptySession, AsyncNestedSession, AsyncOneSession
 
 
 @pytest.fixture(scope="session")
@@ -20,23 +21,23 @@ def event_loop() -> AbstractEventLoop:
 
 
 @pytest.fixture(scope="session")
-def session() -> Generator[latest.Session, None, None]:
+def session() -> Generator[Session, None, None]:
     with open_session() as session:
         yield session
         session.__check__()
 
 
 @pytest_asyncio.fixture(scope="session")
-async def async_session() -> AsyncGenerator[latest.aio.Session, None]:
+async def async_session() -> AsyncGenerator[AsyncSession, None]:
     with open_session(asyncio=True) as session:
         yield session
         await session.__check__()
 
 
 @pytest_asyncio.fixture(scope="session")
-async def empty_session() -> AsyncGenerator[EmptySession, None]:
+async def empty_session() -> AsyncGenerator[AsyncEmptySession, None]:
     interactive = Interactive.open("omc", asynchronous)
-    with EmptySession(interactive) as session:
+    with AsyncEmptySession(interactive) as session:
         assert await session.loadFile(
             resource_filename(__name__, "src/empty.mo")
         )
@@ -45,9 +46,9 @@ async def empty_session() -> AsyncGenerator[EmptySession, None]:
 
 
 @pytest_asyncio.fixture(scope="session")
-async def one_session() -> AsyncGenerator[OneSession, None]:
+async def one_session() -> AsyncGenerator[AsyncOneSession, None]:
     interactive = Interactive.open("omc", asynchronous)
-    with OneSession(interactive) as session:
+    with AsyncOneSession(interactive) as session:
         assert await session.loadFile(
             resource_filename(__name__, "src/one.mo")
         )
@@ -56,9 +57,9 @@ async def one_session() -> AsyncGenerator[OneSession, None]:
 
 
 @pytest_asyncio.fixture(scope="session")
-async def nested_session() -> AsyncGenerator[NestedSession, None]:
+async def nested_session() -> AsyncGenerator[AsyncNestedSession, None]:
     interactive = Interactive.open("omc", asynchronous)
-    with NestedSession(interactive) as session:
+    with AsyncNestedSession(interactive) as session:
         assert await session.loadFile(
             resource_filename(__name__, "src/Nested.mo")
         )