+++ tests/test_parser.py
@@ -9,34 +9,40 @@ import pytest
 from typing_extensions import Annotated
 
 from omc4py import TypeName, VariableName
-from omc4py.latest import OpenModelica
-from omc4py.modelica import alias, enumeration, external, record
+from omc4py.modelica import alias, enumeration, record
 from omc4py.openmodelica import Component
 from omc4py.parser import cast, parse
+from omc4py.v_1_21.OpenModelica.Scripting import (  # NOTE: update to latest
+    SourceInfo,
+)
 
 
-@external(".OneTwo")
 class OneTwo(enumeration):
+    __omc_class__ = TypeName(".OneTwo")
+
     One = 1
     Two = 2
 
 
-@external(".RecordA")
 @dataclass
 class RecordA(record):
+    __omc_class__ = TypeName(".RecordA")
+
     a: int
 
 
-@external(".RecordAB")
 @dataclass
 class RecordAB(record):
+    __omc_class__ = TypeName(".RecordAB")
+
     a: int
     b: int
 
 
-@external(".ScalarRecord")
 @dataclass
 class ScalarRecord(record):
+    __omc_class__ = TypeName(".ScalarRecord")
+
     real: float
     integer: int
     boolean: bool
@@ -46,9 +52,10 @@ class ScalarRecord(record):
     enumeration_: OneTwo
 
 
-@external(".SequenceRecord")
 @dataclass
 class SequenceRecord(record):
+    __omc_class__ = TypeName(".SequenceRecord")
+
     real: Sequence[float]
     integer: Sequence[int]
     boolean: Sequence[bool]
@@ -483,7 +490,7 @@ def test_cast(typ: Any, val: Any, expected: Any) -> None:
             ),
         ),
         (
-            OpenModelica.Scripting.SourceInfo,
+            SourceInfo,
             """\
 record OpenModelica.Scripting.SourceInfo
     filename = "filename",
@@ -494,7 +501,7 @@ record OpenModelica.Scripting.SourceInfo
     columnEnd=4
 end OpenModelica.Scripting.SourceInfo;
 """,
-            OpenModelica.Scripting.SourceInfo(
+            SourceInfo(
                 fileName="filename",
                 readonly=False,
                 lineStart=1,
@@ -504,7 +511,7 @@ end OpenModelica.Scripting.SourceInfo;
             ),
         ),
         (
-            OpenModelica.Scripting.SourceInfo,
+            SourceInfo,
             """\
 record OpenModelica.Scripting.SourceInfo
     fileName = "fileName",
@@ -515,7 +522,7 @@ record OpenModelica.Scripting.SourceInfo
     columnEnd=4
 end OpenModelica.Scripting.SourceInfo;
 """,
-            OpenModelica.Scripting.SourceInfo(
+            SourceInfo(
                 fileName="fileName",
                 readonly=False,
                 lineStart=1,