+++ tests/test_session.py
@@ -6,11 +6,17 @@ import pytest
 
 import omc4py.session
 import omc4py.session.aio
-from omc4py import TypeName, VariableName, latest, open_session
+from omc4py import TypeName, VariableName, open_session
 from omc4py.openmodelica import Component
+from omc4py.v_1_21 import Session  # NOTE: update to latest
 
-from .session import Enum, one
-from .session.aio import EmptySession, NestedSession, OneSession
+from .session import (
+    AsyncEmptySession,
+    AsyncNestedSession,
+    AsyncOneSession,
+    Enum,
+    One,
+)
 
 
 @pytest.mark.dependency()
@@ -21,7 +27,7 @@ def test_open_session() -> None:
 
 
 @pytest.mark.dependency(depends=["test_open_session"])
-def test_loadString(session: latest.Session) -> None:
+def test_loadString(session: Session) -> None:
     assert session.loadString("type MyEnumeration = enumeration(e, n, u, m);")
     assert session.isType("MyEnumeration")
     assert session.isEnumeration("MyEnumeration")
@@ -34,7 +40,7 @@ def test_loadString(session: latest.Session) -> None:
 
 
 @pytest.mark.dependency(depends=["test_open_session"])
-def test_getClassNames(session: latest.Session) -> None:
+def test_getClassNames(session: Session) -> None:
     assert session.loadString(
         """
 package Test_getClassNames
@@ -103,7 +109,7 @@ type B = enumeration(b);
 
 
 @pytest.mark.dependency(depends=["test_open_session"])
-def test_getComponents(session: latest.Session) -> None:
+def test_getComponents(session: Session) -> None:
     assert session.loadString(
         """
 class Test_getComponents
@@ -223,11 +229,11 @@ end Test_getComponents;
 
 
 @pytest.mark.dependency(depends=["test_open_session"])
-def test_OpenModelica(session: latest.Session) -> None:
+def test_OpenModelica(session: Session) -> None:
     assert session.isPackage("OpenModelica")
 
 
-def test_getMessagesStringInternal(session: latest.Session) -> None:
+def test_getMessagesStringInternal(session: Session) -> None:
     session.getMessagesStringInternal()
     for name in ["XXX", "YYY", "ZZZ"]:
         session.__omc_interactive__.evaluate(name)
@@ -299,25 +305,25 @@ def _check_components(components: list[Component]) -> None:
 
 
 @pytest.mark.asyncio
-async def test_empty_session(empty_session: EmptySession) -> None:
+async def test_empty_session(empty_session: AsyncEmptySession) -> None:
     s = empty_session
-    assert await s.empty() is None  # type: ignore
+    await s.empty()
 
 
 @pytest.mark.asyncio
-async def test_one(one_session: OneSession) -> None:
+async def test_one(one_session: AsyncOneSession) -> None:
     s = one_session
     result = await s.one()
-    assert isinstance(result, one)
+    assert isinstance(result, One)
     assert result == (1.0, 1, True, "1", Enum.One)
-    assert result == one(
+    assert result == One(
         real=1.0, integer=1, boolean=True, string="1", enum=Enum.One
     )
 
 
 @pytest.mark.asyncio
-async def test_nested(nested_session: NestedSession) -> None:
+async def test_nested(nested_session: AsyncNestedSession) -> None:
     s = nested_session
-    assert await s.level_1() == await s.Nested.level()
-    assert await s.level_2() == await s.Nested.Nested.level()
-    assert await s.level_3() == await s.Nested.Nested.Nested.level()
+    assert 1 == await s.Nested.level()
+    assert 2 == await s.Nested.Nested.level()
+    assert 3 == await s.Nested.Nested.Nested.level()